Introduce integration test network sessions to parallelize integration test subtests. (#84)
* Migrate endow to big.int

* Implement integrationi test sessions

This change enables parallel execution of regions sharing a single instance of a  integration test network

* Parallelize Set code tx test

Migrate set code transaction to use integration test sessions, and parallel execution of sub-tests.

* Parallelize self destruct test

Migrate test to use integration test sessions, and parallel execution of sub-tests.

* Remove create account with max balance

Maximum balance is difficult to define, since
there must exist another account with an even larger balance to endow it.
Instead of changing the name for this function and try to work the term, just use a more direct approach.

* Rename session default account getter.

* fixup! Implement integrationi test sessions